fix: renaming sgtk to mimos
